+++ Packages/com.cdm.figma.ui/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2023-07-30
+### Breaking Changes
+- `enableTruncate` changed to `overflowMode`
+- Word wrapping is always enabled (default behaviour in Figma)
+- Ellipsis is used if `TypeStyle.textTruncation` is `TextTruncation.Ending`
+
+### Changes
+- Added `minWidth` and `minHeight` to `LayoutElement` component if they are defines in Figma 
+- `TextCase` is importing as `StylePropertyFontStyle`
+
 ## [1.6.5] - 2023-07-23
 ### Changes
 - Component converters made reusable by using inheritance